--- OLD/listefonctionnalités.txt
@@ -1,19 +1,0 @@
-FETCH Infos User + Write BDD
-FETCH Infos @everyone + Write BDD 
-Update Infos BDD
-FETCH Infos Server + Write BDD
-
-On Join = Message 
-
-BACKUP BDD ALT FILE
-
-COMMANDS LIST EMBED
-
-PING
-
-La Base de données stockes les infos serveur + les infos user avec l'ID comme référence. 
-
-if ctx.author.id == AUTHORIZED_USER_ID:
-        await ctx.send("Commande exécutée avec succès !")
-    else:
-        await ctx.send("Désolé, tu n'es pas autorisé à exécuter cette commande.")
