+++ commands/templates.py
@@ -1,0 +1,7 @@
+from discord.ext import commands
+import discord
+
+from data.datafunction import fetch_server_data  # exemple d'importations de fonctions externes
+
+#def server_commands(bot):
+        ## CODE
