--- data/data.json
@@ -1,24 +1,0 @@
-{
-    "servers": {
-        "1308806097435623434": {
-            "id": 1308806097435623434,
-            "name": "Entre potes au top",
-            "owner": "ii_ytforest_ii",
-            "member_count": 10,
-            "created_at": "2024-11-20 14:48:10.070000+00:00",
-            "text_channels": 4,
-            "voice_channels": 2,
-            "roles_count": 5,
-            "preferred_locale": [
-                "american_english",
-                "en-US"
-            ],
-            "icon_url": "https://cdn.discordapp.com/icons/1308806097435623434/188313259525f483d55d9579f43ed4a4.png?size=1024",
-            "verification_level": "none",
-            "security_settings": {
-                "explicit_content_filter": "disabled",
-                "default_message_notifications": "all_messages"
-            }
-        }
-    }
-}
