+++ variables.py
@@ -1,0 +1,6 @@
+prefix = ';'
+WELCOME_CHANNEL_ID = 1347277296788177037 # <-- Sur 'SERVEUR DE TEST' ACTUELLEMENT
+DATA_FILE = "data\database.json" # <-- PATH DE LA BASE DE DONNEES
+COULEUR_EMBED_INFO = 0xff1100
+AUTHORIZED_USER_ID = 724954095042953246
+
